% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVCF.R
\name{loadVCF}
\alias{loadVCF}
\title{Load a VCF object}
\usage{
loadVCF(vcf.info, project)
}
\arguments{
\item{vcf.info}{Named list of metadata describing a \linkS4class{VCF} object.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}
}
\value{
A \linkS4class{VCF} object, expanded or collapsed depending on the specification in \code{vcf.info}.
}
\description{
Load a \linkS4class{VCF} object from its staged file contents.
}
\details{
This function assumes that the files were generated by the \code{\link{stageObject}} method for \linkS4class{VCF} subclasses.
It does \emph{not} load a VCF file!
Use \code{\link{scanVcf}} instead if you want to create a \linkS4class{VCF} object from a VCF file.
}
\examples{
fl <- system.file("extdata", "structural.vcf", package="VariantAnnotation")
vcf <- readVcf(fl, genome="hg19")

tmp <- tempfile()
dir.create(tmp)
info <- stageObject(vcf, dir=tmp, path="experiment-1")
loadVCF(info, tmp)

}
\author{
Aaron Lun
}
