% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVCFHeader.R
\name{loadVCFHeader}
\alias{loadVCFHeader}
\title{Load a VCF header}
\usage{
loadVCFHeader(info, project)
}
\arguments{
\item{info}{Named list of metadata for a VCF file.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}
}
\value{
A \linkS4class{VCFHeader} object.
}
\description{
Load the headers of a VCF file into a \linkS4class{VCFHeader} object.
}
\details{
As the name suggests, this only loads the headers of the VCF file.
To load all contents into memory, use \code{\link{scanVcf}} instead.
}
\examples{
fl <- system.file("extdata", "structural.vcf", package="VariantAnnotation")
hdr <- scanVcfHeader(fl)

tmp <- tempfile()
dir.create(tmp)
info <- stageObject(hdr, dir=tmp, path="header")
loadVCFHeader(info, tmp)

}
\author{
Aaron Lun
}
