% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF.R
\name{readVCF}
\alias{readVCF}
\alias{loadVCF}
\alias{loadVCFHeader}
\title{Read a VCF object from disk}
\usage{
readVCF(path, metadata, ...)
}
\arguments{
\item{path}{String containing a path to a directory, usually generated by the \code{\link[alabaster.base]{saveObject}} method for \link[VariantAnnotation]{VCF} objects.}

\item{metadata}{Named list of metadata for this object, see \code{\link[alabaster.base]{readObjectFile}} for details.}

\item{...}{Further arguments passed to internal \code{\link[alabaster.base]{altReadObject}} calls.}
}
\value{
A \link[VariantAnnotation]{VCF} object.
}
\description{
Read a \link[VariantAnnotation]{VCF} object from its on-disk representation.
}
\examples{
fl <- system.file("extdata", "structural.vcf", package="VariantAnnotation")
vcf <- readVcf(fl)

tmp <- tempfile()
saveObject(vcf, tmp)
readObject(tmp)

}
\seealso{
\code{\link{saveObject,VCF-method}}, to save VCF objects to disk.
}
\author{
Aaron Lun
}
