% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageVCF.R
\name{stageObject,VCF-method}
\alias{stageObject,VCF-method}
\title{Stage a VCF object}
\usage{
\S4method{stageObject}{VCF}(x, dir, path, child = FALSE, ...)
}
\arguments{
\item{x}{Any instance of a \linkS4class{VCF} class or one of its subclasses.}

\item{dir}{String containing the path to the staging directory.}

\item{path}{String containing a prefix of the relative path inside \code{dir} where \code{x} is to be saved.
The actual path used to save \code{x} may include additional components, see Details.}

\item{child}{Logical scalar indicating whether \code{x} is a child of a larger object.}

\item{...}{Further arguments to pass to \code{\link{stageObject,RangedSummarizedExperiment-method}}.}
}
\value{
The contents of \code{x} are saved to file inside \code{path}.
A named list containing metadata is returned.
}
\description{
Save the contents of a \linkS4class{VCF} object to file.
}
\details{
Note that we do \emph{not} save the contents of \code{x} in VCF format.
Rather, we re-use the existing machinery for staging SummarizedExperiments from the \pkg{alabaster.se}.
This is more amenable for random access by feature/sample and ensures that we are consistent with the expectations of the parent class.
Applications requiring actual VCF files can instead use \code{\link{writeVcf}} to generate them from \code{x}.
}
\examples{
fl <- system.file("extdata", "structural.vcf", package="VariantAnnotation")
vcf <- readVcf(fl, genome="hg19")

tmp <- tempfile()
dir.create(tmp)
stageObject(vcf, dir=tmp, path="experiment-1")

}
\author{
Aaron Lun
}
