% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageVCFHeader.R
\name{stageObject,VCFHeader-method}
\alias{stageObject,VCFHeader-method}
\title{Stage the VCF headers}
\usage{
\S4method{stageObject}{VCFHeader}(x, dir, path, child = FALSE)
}
\arguments{
\item{x}{A \linkS4class{VCFHeader} object.}

\item{dir}{String containing the path to the staging directory.}

\item{path}{String containing a prefix of the relative path inside \code{dir} where \code{x} is to be saved.
The actual path used to save \code{x} may include additional components, see Details.}

\item{child}{Logical scalar indicating whether \code{x} is a child of a larger object.}
}
\value{
The contents of \code{x} are saved to file inside \code{path}.
A named list containing metadata is returned.
}
\description{
Save the contents of a \linkS4class{VCFHeader} object to file.
This is formatted as a valid VCF file that lacks any entries.
}
\examples{
fl <- system.file("extdata", "structural.vcf", package="VariantAnnotation")
hdr <- scanVcfHeader(fl)

tmp <- tempfile()
dir.create(tmp)
stageObject(hdr, dir=tmp, path="headers")

}
\author{
Aaron Lun
}
